+++ zgw_consumers/admin.py
@@ -1,10 +1,13 @@
 from django.contrib import admin
+from django.db import models
+from django.http import HttpRequest
 
 from privates.admin import PrivateMediaMixin
 from solo.admin import SingletonModelAdmin
 
-from .admin_fields import get_nlx_field, get_zaaktype_field
+from .admin_fields import get_nlx_field
 from .models.services import NLXConfig, Service
+from .settings import get_setting
 from .widgets import NoDownloadPrivateFileWidget
 
 
@@ -13,6 +16,14 @@ class ServiceAdmin(admin.ModelAdmin):
     list_display = ("label", "api_type", "api_root", "nlx", "auth_type")
     list_filter = ("api_type", "auth_type")
     search_fields = ("label", "api_root", "nlx", "uuid")
+
+    def get_fields(self, request: HttpRequest, obj: models.Model | None = None):
+        fields = super().get_fields(request, obj=obj)
+        if get_setting("ZGW_CONSUMERS_IGNORE_OAS_FIELDS"):
+            assert isinstance(fields, list)
+            fields.remove("oas")
+            fields.remove("oas_file")
+        return fields
 
     def formfield_for_dbfield(self, db_field, request, **kwargs):
         if db_field.name == "nlx":
@@ -25,12 +36,3 @@ class ServiceAdmin(admin.ModelAdmin):
 class NLXConfigAdmin(PrivateMediaMixin, SingletonModelAdmin):
     private_media_fields = ("certificate", "certificate_key")
     private_media_file_widget = NoDownloadPrivateFileWidget
-
-
-class ListZaaktypenMixin:
-    zaaktype_fields = ()
-
-    def formfield_for_dbfield(self, db_field, request, **kwargs):
-        if db_field.name in self.zaaktype_fields:
-            return get_zaaktype_field(db_field, request, **kwargs)
-        return super().formfield_for_dbfield(db_field, request, **kwargs)